+++ src/test/java/starter/Genre/GenreStepDef.java
@@ -12,7 +12,7 @@ import java.io.File;
 
 import static org.hamcrest.Matchers.equalTo;
 
-public class GETGenreStepDef {
+public class GenreStepDef {
 
     @Steps
     GenreAPI genreAPI;
@@ -104,5 +104,75 @@ public class GETGenreStepDef {
     @And("validate invalid id mentors all genres JSON Schema")
     public void validateInvalidIdMentorsAllGenresJSONSchema() {File json = new File(Constants.JSON_SCHEMA_GENRE+"GETmentorallgenreswithinvalididJSONSchema.json");
         SerenityRest.and().body(JsonSchemaValidator.matchesJsonSchema(json));
+    }
+    //////////////////////////////////////////////////////////////////////////////////////////////////////////////
+    // post mentor genress
+    @When("Post mentor genress with valid request body")
+    public void postmentorgenress(){
+        File json = new File(Constants.REQ_BODY_GENRE+"/POSTmentorgenrewithvalidrequestbody.json");
+        genreAPI.postmentorgenres(json);
+    }
+
+    @And("request mentor genress with valid request body")
+    public void requestMentorGenressWithValidRequestBody() {SerenityRest.when().post(GenreAPI.POST_GENRES_MENTOR_VALID_REQUEST);
+
+    }
+
+    @And("Validate response body mentor genress message {string}")
+    public void validateResponseBodyMentorGenressMessage(String message) {SerenityRest.and().body(MusiclabResponses.MESSAGE,equalTo(message));
+    }
+
+    @And("Validate message JSON Schema mentor genress")
+    public void validateMessageJSONSchemaMentorGenress() {File json = new File(Constants.JSON_SCHEMA_GENRE+"POSTmentorgenresJSONSchema.json");
+        SerenityRest.and().body(JsonSchemaValidator.matchesJsonSchema(json));
+    }
+    // invalid request body
+    @When("Post mentor genress with invalid request body")
+    public void postMentorGenressWithInvalidRequestBody() {File json = new File(Constants.REQ_BODY_GENRE+"/POSTmentorgenrewithinvalidrequestbody.json");
+        genreAPI.postmentorgenres(json);
+    }
+
+    @And("request mentor genress with invalid request body")
+    public void requestMentorGenressWithInvalidRequestBody() {SerenityRest.when().post(GenreAPI.POST_GENRES_MENTOR_VALID_REQUEST);
+    }
+
+    @And("Validate response invalid body mentor genress message {string}")
+    public void validateResponseInvalidBodyMentorGenressMessage(String message) {SerenityRest.and().body(MusiclabResponses.MESSAGE,equalTo(message));
+    }
+
+    @And("Validate message JSON Schema invalid mentor genress")
+    public void validateMessageJSONSchemaInvalidMentorGenress() {File json = new File(Constants.JSON_SCHEMA_GENRE+"POSTinvalidmentorgenresJSONSchema.json");
+        SerenityRest.and().body(JsonSchemaValidator.matchesJsonSchema(json));
+    }
+    ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+    // valid data delete
+    @When("Delete mentor with valid paramater id {int}")
+    public void deletementorwithvalidparamaterid(int id){
+        genreAPI.postmentorgenresid(id);
+    }
+
+    @And("request for delete mentor genre")
+    public void requestForDeleteMentorGenre() {SerenityRest.when().delete(GenreAPI.DELETE_ALL_MENTOR_WITH_ID);
+    }
+
+    @And("Validate response body delete message should be {string}")
+    public void validateResponseBodyDeleteMessageShouldBe(String message) {SerenityRest.and().body(MusiclabResponses.MESSAGE,equalTo(message));
+    }
+
+    @And("Validate message delete JSON Schema")
+    public void validateMessageDeleteJSONSchema() {File json = new File(Constants.JSON_SCHEMA_GENRE+"DELETEallmentorJSONSchema.json");
+        SerenityRest.and().body(JsonSchemaValidator.matchesJsonSchema(json));
+    }
+    // invalid id delete
+    @When("Delete mentor with invalid paramater id {}")
+    public void deleteMentorWithInvalidParamater(int id) {genreAPI.postmentorgenresid(id);
+    }
+
+    @And("request for delete mentor genre with invalid params")
+    public void requestForDeleteMentorGenreWithInvalidParams() {SerenityRest.when().delete(GenreAPI.DELETE_ALL_MENTOR_WITH_ID);
+    }
+
+    @And("Validate message invalid delete JSON Schema")
+    public void validateMessageInvalidDeleteJSONSchema() {File json = new File(Constants.JSON_SCHEMA_GENRE+"DELETEallmentorJSONSchema.json");
     }
 }
