+++ src/test/resources/features/Student/GetStudentprofile.feature
@@ -1,6 +1,6 @@
 Feature: Get Studen Profile
 
-  @Tugas @positivecase
+  @Musiclab @PositiveCase @GetALL @Student
   Scenario: Get show student profile with valid id
     Given User login student with account
     When show student profile with valid params
@@ -9,7 +9,7 @@ Feature: Get Studen Profile
     And Validate response body for message should be "success show student profile"
     And Validate get student profile JSON Schema
 
-@Tugas @Negativecase
+  @Musiclab @NegativeCase @GetALL @Student
   Scenario: Get show student profile with invalid Method
     Given User login student with account
     When show student profile with invalid method
@@ -18,7 +18,7 @@ Feature: Get Studen Profile
     And Validate response invalid method body for message should be "Method Not Allowed"
     And Validate get student profile invalid method JSON Schema
 
-@Tugas @Negativecase
+  @Musiclab @NegativeCase @GetALL @Student
   Scenario: Get show studen profile with invalid params
     Given User login student with account
     When show student profile with invalid params
