+++ src/test/resources/features/Student/PUTStudentprofle.feature
@@ -1,6 +1,6 @@
 Feature: PUT Student Profile
 
-  @Tugas
+  @Musiclab @PositiveCase @PutALL @Student
   Scenario: Put edit student profile with valid request body
     Given User login student with account
     When put edit student profile with valid request body
@@ -10,7 +10,7 @@ Feature: PUT Student Profile
     And Validate message respon body JSON Schema
 
 
-    @Tugas
+  @Musiclab @NegativeCase @PutALL @Student
     Scenario: Put edit student profile with empty name on request body
       Given User login student with account
       When put edit student profile with empty name on request body
@@ -18,7 +18,7 @@ Feature: PUT Student Profile
       Then Status code should be 400 Bad Request
       And Validate respon body "validate: Key: 'Core.Name' Error:Field validation for 'Name' failed on the 'required' tag"
 
-  @Tugas
+  @Musiclab @NegativeCase @PutALL @Student
   Scenario: Put Edit student profil with invalid params
     Given User login student with account
     When put edit student profile with invalid params
@@ -26,7 +26,7 @@ Feature: PUT Student Profile
     Then Status code should be 404 Not Found
     And Validate response body for message "Not Found"
 
-    @Tugas
+  @Musiclab @NegativeCase @PutALL @Student
     Scenario: Put Edit student profil with empty request body
       Given User login student with account
       When put edit studet profile with empty request body
