+++ src/test/resources/features/Transaction/GETtransactions.feature
@@ -1,5 +1,5 @@
 Feature: Get mentor transaction
-
+  @Musiclab @PositiveCase @GetALL @Mentor
   Scenario Outline: Get mentor transactions with valid parameter page
     Given User login mentor with account
     And Get mentor transactions with valid parameter <page> <limit>
@@ -10,7 +10,7 @@ Feature: Get mentor transaction
       | 1    | 100    |
       | 5    | 20     |
 
-
+  @Musiclab @NegativeCase @GetALL @Mentor
   Scenario Outline: Get mentor transactions with invalid parameter page and limit
     Given User login mentor with account
     And Get mentor transactions with invalid parameter <page> <limit>
@@ -21,6 +21,7 @@ Feature: Get mentor transaction
       | @   | JHO    |
       | @@    | JU     |
 
+  @Musiclab @NegativeCase @GetALL @Mentor
   Scenario Outline: Get mentor transactions with empty value
     Given User login mentor with account
     And Get mentor transactions with empty value parameter <page> <limit>
@@ -29,7 +30,10 @@ Feature: Get mentor transaction
     Examples:
       | page | limit |
       |      |        |
+
+
 # Student Transaction
+  @Musiclab @PositiveCase @GetALL @Student
   Scenario Outline: Get students transactions with valid params
     Given User login student with account
     And Get students transactions with valid parameter <page> <limit>
@@ -40,6 +44,7 @@ Feature: Get mentor transaction
       |  1    |   100     |
       |  5    |   50     |
 
+  @Musiclab @NegativeCase @GetALL @Student
   Scenario Outline: Get students transactions with invalid params
     Given User login student with account
     And Get students transactions with invalid parameter <page> <limit>
@@ -49,7 +54,7 @@ Feature: Get mentor transaction
       | page | limit |
       |  ee   |   bb     |
       |  xx    |   PUBG    |
-
+  @Musiclab @NegativeCase @GetALL @Student
   Scenario Outline: Get students transactions with empty value params
     Given User login student with account
     And Get students transactions with empty value params <page> <limit>
