+++ src/test/resources/features/Transaction/POSTmakeTransaction.feature
@@ -1,6 +1,6 @@
 Feature: POST make transaction mentor and student
 
-  @Tugas
+  @Musiclab @PositiveCase @PostALL @Student
   Scenario: Post Make transaction with valid data
     Given User login student with account
     When Post make transaction with valid request body
@@ -9,7 +9,7 @@ Feature: POST make transaction mentor and student
     And Validate response body message make transaction "success make transaction"
     And Validate JSON Schema make transaction
 
-  @Tugas
+  @Musiclab @NegativeCase @PostALL @Student
   Scenario: Post Make transaction with empty class id
     Given User login student with account
     When Post make transaction with empty class id request body
@@ -18,7 +18,7 @@ Feature: POST make transaction mentor and student
     And Validate response body message empty class id make transaction "input format incorrect"
     And Validate JSON Schema make transaction empty class id
 
-  @Tugas
+  @Musiclab @NegativeCase @PostALL @Student
   Scenario: Post Make transaction with schedule id
     Given User login student with account
     When Post make transaction with schedule id request body
@@ -27,7 +27,7 @@ Feature: POST make transaction mentor and student
     And Validate response body message empty class id make transaction "input format incorrect"
     And Validate JSON Schema make transaction empty schedule id
 
-  @Tugas
+  @Musiclab @NegativeCase @PostALL @Student
   Scenario: Post Make transaction with invalid start date
     Given User login student with account
     When Post make transaction with invalid start date request body
@@ -36,7 +36,7 @@ Feature: POST make transaction mentor and student
     And Validate response body message make transaction "minimum start date input is today"
     And Validate JSON Schema make transaction invalid start date
 
-  @Tugas
+  @Musiclab @NegativeCase @PostALL @Student
   Scenario: Post Make transaction with invalid parameter
     Given User login student with account
     When Post make transaction with invalid paramter
@@ -45,7 +45,7 @@ Feature: POST make transaction mentor and student
     And Validate response body message make transaction "Not Found"
     And Validate JSON Schema make transaction with invalid parameter
 
-  @Tugas
+  @Musiclab @NegativeCase @PostALL @Student
   Scenario: Post Make transaction with empty body request
     Given User login student with account
     When Post make transaction with empty body request
